improving data type in calculator

+++ projects/js-calculator/script.js
@@ -1,5 +1,5 @@
-const firstNumber = parseInt(prompt("Enter your first number"));
-const secondNumber = parseInt(prompt("Enter your second number"));
+const firstNumber = parseFloat(prompt("Enter your first number"));
+const secondNumber = parseFloat(prompt("Enter your second number"));
 
 alert(`The addition of ${firstNumber} and ${secondNumber} is equal to ${firstNumber + secondNumber}
 The substraction of ${firstNumber} and ${secondNumber} is equal to ${firstNumber - secondNumber}
